+++ include/morphology/logical_or_binary.h
@@ -3,7 +3,7 @@
 
 //kernel for image logical_or
 template <typename dtype>
-void logical_or(dtype* deviceImage, dtype* deviceOutput, const int size,
+void logical_or(dtype* deviceImage, dtype* deviceOutput, const size_t size,
                        const int flag_verbose);
 
 // Slide kernel and erosion operation over all input image pixels
@@ -13,6 +13,6 @@ void logical_or_on_device(dtype* hostImage, dtype* hostOutput, const int xsize,
 
 // Slide kernel and erosion operation over all input image pixels
 template <typename dtype>
-void logical_or_on_host(dtype* hostImage, dtype* hostOutput, const int size);
+void logical_or_on_host(dtype* hostImage, dtype* hostOutput, const size_t size);
 
 #endif  // LOGICAL_OR_H
